(LIDAR-212) Fix Comments
