+++ cmd/puppetdb/pdb.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"net/url"
 	"os"
-	"regexp"
-	"strconv"
 	"strings"
 
 	prompt "github.com/c-bata/go-prompt"
@@ -52,7 +50,7 @@ func executor(in string) {
 	in = strings.TrimSpace(in)
 
 	// Parse the input and extract the API call + query
-	var api, query, pagination = parseInput(in)
+	var api, query, pagination = cli.ParseInput(in)
 	// If a api has been selected, then execute it with the provided query
 	// the command should be recorded in history and the response printed to
 	// stdout
@@ -73,7 +71,7 @@ func execute(api string, query string, pagination puppetdb.Pagination) {
 	switch api {
 	case "nodes":
 		fmt.Printf("Nodes")
-		data, err = client.Nodes(query, pagination)
+		data, err = client.Nodes(query, &pagination)
 	case "facts":
 		data, err = client.Facts(query)
 	case "inventory":
@@ -89,81 +87,6 @@ func execute(api string, query string, pagination puppetdb.Pagination) {
 		return
 	}
 	cli.PrintString(data)
-}
-
-func extractString(str, start, end string) (result string) {
-	s := strings.Index(str, start)
-	s--
-	if s == -1 {
-		return
-	}
-	s += len(start)
-	e := strings.LastIndex(str, end)
-	e++
-	if e == -1 {
-		return
-	}
-	return str[s:e]
-}
-
-func parseInput(command string) (string, string, puppetdb.Pagination) {
-
-	//Split Query Api and Pagination Up
-	checkForQuery, err := regexp.Match(`[\\w+]`, []byte(command))
-	var query string
-	if checkForQuery {
-		query = extractString(command, "[", "]")
-		if query == "[]" {
-			query = ""
-		}
-	}
-	if err != nil {
-		fmt.Println("No query parameters detected", err)
-	}
-
-	querylessCommand := strings.Replace(command, query, "", 1)
-
-	blocks := strings.Split(querylessCommand, " ")
-	api := blocks[0]
-
-	if api == "exit" {
-		fmt.Println("Bye!")
-		os.Exit(0)
-	}
-
-	var rex = regexp.MustCompile(`(\w+)=(\w+)`) //nolint
-
-	options := rex.FindAllString(querylessCommand, -1)
-
-	pagination := puppetdb.Pagination{}
-
-	for _, option := range options {
-
-		re := regexp.MustCompile(`Limit=(\d*)`)
-		limit := re.FindStringSubmatch(option)
-
-		re2 := regexp.MustCompile(`Offset=(\d*)`)
-		offset := re2.FindStringSubmatch(option)
-
-		re3 := regexp.MustCompile(`Include_total=(\d*)`)
-		total := re3.FindStringSubmatch(option)
-
-		if len(limit) >= 1 {
-			pagination.Limit, err = strconv.Atoi(limit[1])
-		}
-		if len(offset) >= 1 {
-
-			pagination.Offset, err = strconv.Atoi(offset[1])
-		}
-		if len(total) >= 1 {
-			fmt.Println("Include_total is currently not implemented")
-		}
-		if err != nil {
-			fmt.Println("err: " + err.Error())
-		}
-	}
-
-	return api, query, pagination
 }
 
 func completer(in prompt.Document) []prompt.Suggest {
