+++ pkg/puppetdb/nodes.go
@@ -1,33 +1,15 @@
 package puppetdb
 
-import (
-	"fmt"
-	"strconv"
-)
-
-func (p Pagination) toParams() map[string]string {
-	params := map[string]string{}
-	if p.Limit > 0 {
-		params["limit"] = strconv.Itoa(p.Limit)
-	}
-	if p.Offset > 0 {
-		params["offset"] = strconv.Itoa(p.Offset)
-	}
-
-	if p.IncludeTotal {
-		params["include_total"] = strconv.FormatBool(p.IncludeTotal)
-	}
-	return params
-}
+import "fmt"
 
 // Nodes will return all nodes matching the given query. Deactivated and expired nodes aren’t included in the response.
-func (c *Client) Nodes(query string, pagination Pagination) ([]Node, error) {
+func (c *Client) Nodes(query string, pagination *Pagination) ([]Node, error) {
 	payload := []Node{}
 	req := c.resty.R().SetResult(&payload)
 	if query != "" {
 		req.SetQueryParam("query", query)
 	}
-	if pagination.toParams() != nil {
+	if pagination != nil {
 		req.SetQueryParams(pagination.toParams())
 	}
 
@@ -59,11 +41,4 @@ type Node struct {
 	CatalogTimestamp             string      `json:"catalog_timestamp"`
 	LatestReportJobID            string      `json:"latest_report_job_id"`
 	LatestReportStatus           string      `json:"latest_report_status"`
-}
-
-// Pagination is a filter to be used when paginating
-type Pagination struct {
-	Limit        int
-	Offset       int
-	IncludeTotal bool
 }
